Cleanup initializers & static properties

+++ Sources/Swift2D/Point+CoreGraphics.swift
@@ -7,13 +7,13 @@ import Foundation
 public extension Point {
     /// Initialize a `Point` using the provided `CGPoint` values.
     init(_ point: CGPoint) {
-        self.init(x: Double(point.x), y: Double(point.y))
+        self.init(x: point.x, y: point.y)
     }
 }
 
 public extension CGPoint {
     /// Initialize a `CPPoint` using the provided `Point` values.
     init(_ point: Point) {
-        self.init(x: CGFloat(point.x), y: CGFloat(point.y))
+        self.init(x: point.x, y: point.y)
     }
 }
