+++ Sources/Swift2D/Size+CoreGraphics.swift
@@ -7,13 +7,13 @@ import Foundation
 public extension Size {
     /// Initialize a `Size` using the provided `CGSize` values.
     init(_ size: CGSize) {
-        self.init(width: Double(size.width), height: Double(size.height))
+        self.init(width: size.width, height: size.height)
     }
 }
 
 public extension CGSize {
     /// Initialize a `CGSize` using the provided `Size` values.
     init(_ size: Size) {
-        self.init(width: CGFloat(size.width), height: CGFloat(size.height))
+        self.init(width: size.width, height: size.height)
     }
 }
